+++ changelog.md
@@ -11,7 +11,7 @@
 
 ## API Changes
 
-- ImageInterface::blendTransparency() was renamed to ImageInterface::background()
+- ImageInterface::blendTransparency() was renamed to ImageInterface::background() - Signature changed
 - ImageInterface::setBlendingColor() was renamed to ImageInterface::setBackgroundColor()
 - ImageInterface::blendingColor() was renamed to ImageInterface::backgroundColor()
 - Changed default value for `background` to `null` in ImageInterface::rotate()